Revert "Revert "fixing a bunch of warnings in visual studio 2010""
This reverts commit bfbfd8be0c0f8f2037a4d1e3b1785a669ffab432.

+++ Source/MediaStorageAndFileFormat/gdcmSpacing.cxx
@@ -86,7 +86,7 @@ double frap(double frac[2], double startx, double maxden = 10 )
   const double error = startx - ((double) m[0][0] / (double) m[1][0]);
 
   /* now try other possibility */
-  ai = (maxden - m[1][1]) / m[1][0];
+  ai = ((long)maxden - m[1][1]) / m[1][0];
   m[0][0] = m[0][0] * ai + m[0][1];
   m[1][0] = m[1][0] * ai + m[1][1];
   //printf("%ld/%ld, error = %e\n", m[0][0], m[1][0],
